Move TYPED_TEST_SUITE fix to the proper place
