+++ core/test/dataset_binary_arithmetic_test.cpp
@@ -2,14 +2,13 @@
 // Copyright (c) 2019 Scipp contributors (https://github.com/scipp)
 #include <initializer_list>
 
-#include "test_macros.h"
 #include "test_operations.h"
 #include <gtest/gtest-matchers.h>
 #include <gtest/gtest.h>
 
-#include "scipp/common/fix_typed_test_suite_warnings.h"
 #include "scipp/core/dataset.h"
 #include "scipp/core/dimensions.h"
+#include "test_macros.h"
 
 #include "dataset_test_common.h"
 #include "make_sparse.h"
