+++ core/test/dataset_operations_test.cpp
@@ -1,12 +1,11 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 // Copyright (c) 2019 Scipp contributors (https://github.com/scipp)
-#include "test_macros.h"
 #include <gtest/gtest-matchers.h>
 #include <gtest/gtest.h>
 
 #include "dataset_test_common.h"
-#include "scipp/common/fix_typed_test_suite_warnings.h"
 #include "scipp/core/dataset.h"
+#include "test_macros.h"
 
 using namespace scipp;
 using namespace scipp::core;
