+++ test/fix_typed_test_suite_warnings.h
